Implement reviews (#127)
Guests and Interessent are allowed to send a review some days after the events started
Creates a unique token for each reservation request and sends in invitation to each requester. 
A star - rating (1 of 5), a comment and a optional name can be sent as a review message.

Each location then has a new section to show the incoming reviews and an average rating.